requirements.txt
# SmartMoneyAI v3 - Зависимости

# API и WebSocket
requests>=2.31.0
websockets>=12.0
ccxt>=4.1.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Technical Analysis
ta-lib>=0.4.28
pandas-ta>=0.3.14b0

# Telegram Bot
python-telegram-bot>=20.7

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio>=3.4.3

# Logging
colorlog>=6.8.0

# Testing (опционально)
pytest>=7.4.0
pytest-asyncio>=0.21.0

